Atualiza código do PS2Doom

+++ source/ps2doom.h
@@ -5,7 +5,8 @@
 
 
 
-int gethostname(char *name, int len);
+//int gethostname(char *name, int len);
+int gethostname(char *name, size_t len);
 u32 inet_addr(const char *cp);
 
 float pow(float a, float b);
